requirements.txt
Django>=1.4.3
django-appconf>=0.4
envoy==0.0.2